HAI-450 HAI-449
Map with projects is separated from HankeDrawer where geometries of
projects are edited.

Only common VectorLayer is used and styles given to it as props.
No need for a separate HankeGeometryLayer.

Features are given into a single VectorLayer, there is no need to map
through features and have them on separate layers at this point.