HAI-1351 Fix title and subtitle for logged in users
Add a welcome text for logged in users. The title of the software should
be different depending on whether the is logged in or not, though I
don't know why.

The difference is whether Haitaton is called 'palvelu' or
'asiointi'/'asiointijärjestelmä'.

The new titles should match the ones in Figma.