HAI-1455 Implement naming of cable report application areas (#257)
User can name application areas and the name is displayed as that
areas tab name. Name can be maximum 30 characters long and if it is
longer than that it will be shortened to 30. If no name is given,
a default name will be used.

geometry field in applicationData is changed to areas field which contains
a list of areas. Each area has a name and geometry fields.