HAI-1456 User can cancel new application (#260)
Implemented functionality that user can cancel new application
if it is possible, meaning it is not yet in Allu or if the status
is still pending.

If cancelling is not possible, cancel button is not displayed.

After successful cancellation, user is taken in to application list
and success notification is shown. If something goes wrong, error message
is shown.