HAI-1523 Prevent deletion of hanke if it has applications being handled in Allu
User can not delete hanke if it has any applications that have proceeded
to being handled in Allu. If it has such applications cancel button is hidden
in hanke view. In a case where user has pressed the cancel button when at that
moment no application is being handled but that changes before user confirms
cancelling, info dialog is displayed to user that deletion is not possible.