HAI-1750 Display loading indicator in place of stepper steps when uploading attachments
When uploading attachments in cable report application form, it was possible
to go to last page of the form by clicking the last step heading in top of the form
and send the application before all attachments had been uploaded. Changed it so
that stepper headings are replaced with loading indicator while attachments are uploaded.