HAI-1843 Pre-create env-config.js on the target image
Give write permissions for it to all users. This should enable the
random user that's running nginx to write the config based on
environment variables.

Remove the explicit delete and touch from the env.sh script. Instead,
overwrite anything in the file with the first write operation the script
makes.