HAI-1586 Change hanke contact data model (#370)
Change hanke contact data model

- Split sub contact name to first name and last name.
- Remove address information from contacts.
- sub contact phone is now mandatory.
- Sub contact form should be always visible (if contact is)
- You can add many sub contacts.

Additional: the ticket does not mention this but ytunnusTaiHetu was changed to just ytunnus. It is visible if contact is not a person.

Note: This commit does not fix / change everything that needs to be changed in yhteystieto page. For example, the styling of the page does not figma, but that task is not part of this task.