HAI-1463 Change frontend to use hanke status returned from backend (#391)
Status of hanke can be either DRAFT, PUBLIC or ENDED and this
information is returned from backend. Changed frontend to keep
track of the backend state.

At this point this affects displaying draft state notification
if hanke status is DRAFT as well as displaying
Add application button in hanke form and hanke view
if hanke status is PUBLIC.

Also removed the saveType field as it's not needed anymore.