HAI-2008 Allow user to name hanke areas (#394)
Added text input for hanke area card in areas page of hanke form,
which allows user to enter name for each hanke area.

Name input has a default value that is Hankealue + the next running number
of the default names of other hanke areas. For example if there are
names Hankealue 1, Hankealue 2 and Hankealue 4, the next default name
would be Hankealue 5. User can then edit the default name if they want to.

---------

Co-authored-by: Topias Heinonen <topias.heinonen@gofore.com>