HAI-2067 Add error messages to file upload component
Show error message to user for each file that fails to be uploaded,
either because the file is a duplicate, fails client side validation,
fails validation in the server or fails because of other server error.

There are three kind of error messages. One for duplicate file,
one for not passing client side or server side validation and one
for other server errors.