HAI Fix regex warning
Sonarcloud has been reporting abot a potential problem with the regex
`/\d+$/`. This can be avoided by using a capture group.

Use a capture group to find the number of the area, if the area name is
a default name. Both matching and finding the number can be done with
the same regex. This also removes the need to have a separate filtering
step.