HAI-2060 Fix homepage links in login error page (#422)
There was a problem in links back to homepage in login error page
that user is directed to after failed login. Links had href /login
instead of /fi, /sv or /en based on language, which was caused
by language being set to login. Fixed that by defining
supportedLngs in i18n init options.