HAI-1811 Fix focus on page change
Focus was not set to top of the page when navigating to a different
page. Fixed that by setting the focus to empty div on top of the page
so that pressing tab sets the focus to Skip to main content link.
Also added aria-live regions to main page headings.