HAI-1776 Add Content-Security-Policy header (#425)
Add a Content-Security-Policy header to the nginx configuration. The
header tells the browser where it should download content from. This is
added security for exploits like XSS.

When building the docker image, copy nginx.conf under /opt. When the
container is started, run a script at startup that substitutes the
environment variables and copies nging.conf to /etc/nginx/nginx.conf.
The script only substitutes the environment variables specified in the
script so it won't overwrite nginx variables.

Also, modify the docker file for building UI images for the local Docker
Compose environment. Copy the files to the builder image selectively, so
only changes in those files invalidate Docker's cache. Separate `yarn
install` from `yarn build`, so that only changes in package.json or
yarn.lock will trigger a new `yarn install`. This means that the longest
step in building the UI image can be cached most of the time when
switching branches etc.