HAI-2145 Fix changing language redirecting user to front page
There was a problem when user was in a route containing dynamic
path segment such as application id and changed UI language.
In such case user was taken to front page instead of staying in
the same page.

Problem was caused because for example key with value of
'hakemus/4' was not found in language files, so fixed this
by replacing the id with ':id', so in this case 'hakemus/:id,
and then placing the actual id back when navigating to changed
language route.