HAI-2153 Create hanke from modal dialog
Hanke is created from modal dialog where name of the hanke
is entered along with user email and phone number. Dialog is
opened from any of the new hanke links in home page, header or
hanke list.

After the hanke is created when user presses the Create
hanke button in the dialog, user is taken to edit hanke
form.

Removed NewHankePage component and the new hanke route.