HAI-2264 Add page to edit user information
Edit page can be accessed by clicking pen icon link in table. The
links are not displayed if user does not have MODIFY_USER permission,
except for the user themselves.

The edit page has all the fields that can be edited, but no Save,
Cancel or Delete buttons as those functions will be implemented later.