HAI-2391 Implement hanke user delete (#605)
Added delete buttons to user management view and user edit view.

When clicking the delete button, it is checked from backend if user can
be deleted. Dialog showing the delete information is opened
and if user can be deleted there are buttons for confirming and
cancelling. If user deletes themselves from hanke they are taken
to hanke list.