HAI-2573 Change all fields required for public hanke to be required (#740)
Changed validation of all required fields for public hanke to be
actually required. However moving to another page in the hanke
form is possible even with missing fields when hanke is in draft
state. When hanke is public, moving to another page is not allowed
if user removes required information.