HAI-2681 Make sure that empty OVT and registry key fields are sent as null in kaivuilmoitus
There was an issue that if OVT and registry key fields in kaivuilmoitus would be left
empty, they would be sent as empty strings to backend, which caused errors.

Made a change that those fields are null by default and if they are
clered they are set as null.

Also fixed validation of rockExcavation and cableReport fields
and set registry key fields of yhteystiedot as required.