HAI-1554 Improve hanke and work area info overlay
Info overlay was displayed only when hovering over hanke or work area
in johtoselvitys and kaivuilmoitus map in areas page. Changed it so
that overlay is also displayed when area is selected.

Also changed overlay to use OpenLayers Overlay to fix it into one
point on top of area.

Changed overlay to match Figma design and added hanke name and area
duration to info.