HAI-2340 Prevent drawing work areas outside hanke areas
In kaivuilmoitus and johtoselvitys drawing outside hanke areas
is prevented. However this restriction does not apply to johtoselvitys
which has generated hanke.

This is implemented by not allowing drawing to start outside of hanke
areas and it also cannot be ended outside. Also modifying work areas
is restricted to hanke areas. Also when drawing tool is active point
geometry is not displayed when cursor is outside hanke areas to indicate
that drawing is not possible.