HAI-215 Show täydennyspyyntö for kaivuilmoitus
Show a notification for a täydennyspyyntö in the application page for a
kaivuilmoitus. The notification is the same as for a johtoselvitys, but
the labels for start and end date are different. Use the application
type in the localization key to show different labels for different
types.

Show the täydennyspyyntö notification for all applications.

Also, make the TaydennyspyyntoNotification test faster by reusing the
item list between iterations.