HAI-131 Fix validation context for haittojenhallinta in hanke and kaivuilmoitus (#1058)
There was a problem with validation context object not updating
when making changes to hanke or work areas. This caused problems
for haittojenhallinta liikennemuoto fields validation which
depend on context given to validation.

Fixed this by making sure context object is always updated when saving
in hanke, kaivuilmoitus and kaivuilmoitus taydennys forms.