HAI-3429 Use Haitaton language in Profiili login
Upgrade HDS libraries to 3.12.0. The ability to add redirection
properties to the login buttons was added in 3.11.0.

Using the new properties, pass the language selected in Haitaton to
Profiili as a query parameter in the auth redirection. This will, at
least most of the time, set the language of the login screen to match
the language selected in Haitaton.

There might be some exceptions since Keycloak considers several factors
when selecting the UI language. The ones most likely to trip us might be
an open session directly to the Profiili service, or if the user has
selected a different language on the Profiili login screen. These should
be very rare, though. While testing, I didn't encounter any problems
even though I tried to do irregular language selection combinations.