HAI-3429 Fix E2E tests after login language change
Update the login sections in Playwright tests. The login page is now in
Finnish since it follows the language of the Haitaton page. E2E tests
use the Finnish page in Haitaton.

Also, add e2e/report to .gitignore, so that git ignores the files
Playwright generates.

Remove the empty readme file from the report directory. I think it was
created to create the report folder when checking out the source from
git. Playwright creates it when the tests are run, so it doesn't need to
be created with an empty file.

Run `yarn lint` on the e2e tests, since it's a real pain not to run it.