HAI Use .env.e2e for local E2E config
When running the Playwright tests locally, read the configuration from
`.env.e2e`. This means there's no need to declare environment variables
to the local environment.

Add an example file for the environment variables as `.env.e2e.example`.
The sensitive values are missing from the file. Add instructions to
README.md for how to set up the configuration and how to run the tests.

Add `.env.e2e` to `.gitignore`, so that the sensitive values are never
added to commits.

Add `e2e/report` to `.gitignore`. Add a description to
`e2e/report/readme` on why the file is needed.

Add a yarn command for running the Playwright tests with default
options: `yarn e2e`.