HAI-221 Don't show validation errors for undetected liikennemuodot
There was a problem in kaivuilmoitus form haittojenhallinta page
that validation errors might be shown for liikennemuodot that
have nuisance index 0 (which are optional and no error should
be shown). This was happening if there were work areas in multiple
hanke areas as nuisance indexes were checked always only on the
first area. Fixed this by finding the area path segment based
on if it includes 'alueet' or 'areas' instead of taking the first
path segment as the first segment for hakemus is 'applicationData'.