HAI Add Allu trigger calls to E2E tests
Add calls to manually trigger Allu updates to E2E tests. The call will
trigger Allu status updates instantly in the backend, so there's no need
to wait up to 60 seconds for the next update. The trigger is available
only in test environments, but we can't and don't want to run the e2e
test against production.

I left the wait loops untouched, in case there's a delay in Allu having
the status update in the status history.