HAI-3522 Fix problems with y-tunnus validation in hanke form
There was a problem that hanke in draft state could not be saved
if y-tunnus was missing for yhteystieto. Also if y-tunnus was invalid
it was not clear for the user what was wrong when trying to go
to another form page as focus was not moved to invalid field and
there was no error notification.

Fix this by triggering the validation through react-hook-form.